ログインしていない状態でもチャットを使えるようにした

+++ front/app/pages/mac/[id].tsx
@@ -1,5 +1,5 @@
 import { GetServerSideProps, NextPage } from 'next';
-import React from 'react';
+import React, { useContext } from 'react';
 import Head from 'next/head';
 import { Header } from 'lib/components/Header';
 import { Footer } from 'lib/components/Footer';
@@ -26,6 +26,9 @@ import 'swiper/css';
 import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 import 'swiper/css/scrollbar';
+import { AuthContext } from 'pages/_app';
+import { AlertState } from 'lib/interfaces';
+import { AlertMessage } from 'lib/components/AlertMessage';
 
 interface Props {
   id: number;
@@ -39,16 +42,30 @@ interface State {
 }
 
 const MacShow: NextPage<Props> = ({ id }) => {
+  const { isSignedIn } = useContext(AuthContext);
   const [state, setState] = useState<State>();
   const [show, setShow] = useState(false);
+  const [alertState, setAlertState] = useState<AlertState>({
+    open: false,
+    type: 'info',
+    message: '',
+  });
 
   const fetchData = async () => {
     const product = await ProductRepository.get(id);
     const mac = await MacRepository.get(product.id);
-    const likes = await LikeRepository.my();
     const imageSwipers = await ImageSwiperRepository.showMac(mac.Id);
-    const likeProductIds = likes.map((l) => l.productId);
-    setState({ product, mac, likeProductIds, imageSwipers });
+    if (isSignedIn) {
+      const likes = await LikeRepository.my();
+      const likeProductIds = likes.map((l) => l.productId);
+      setState({ product, mac, likeProductIds, imageSwipers });
+    }
+    setState({
+      product: product,
+      mac: mac,
+      likeProductIds: [],
+      imageSwipers: imageSwipers,
+    });
   };
 
   useEffect(() => {
@@ -59,18 +76,28 @@ const MacShow: NextPage<Props> = ({ id }) => {
   const { product, mac, likeProductIds, imageSwipers } = state;
 
   const addLikeVButton = async () => {
-    await LikeRepository.create(product.id);
-    setState({ ...state, likeProductIds: [...likeProductIds, product.id] });
+    if (isSignedIn) {
+      await LikeRepository.create(product.id);
+      setState({ ...state, likeProductIds: [...likeProductIds, product.id] });
+    } else {
+      setAlertState({
+        open: true,
+        type: 'warning',
+        message: 'ログインすることで商品をお気に入りに登録することができます',
+      });
+    }
   };
 
   const deleteLikeButton = async () => {
-    await LikeRepository.delete(product.id);
-    setState({
-      ...state,
-      likeProductIds: likeProductIds.filter(
-        (productId) => productId !== product.id
-      ),
-    });
+    if (isSignedIn) {
+      await LikeRepository.delete(product.id);
+      setState({
+        ...state,
+        likeProductIds: likeProductIds.filter(
+          (productId) => productId !== product.id
+        ),
+      });
+    }
   };
 
   return (
@@ -143,7 +170,7 @@ const MacShow: NextPage<Props> = ({ id }) => {
             </div>
             <div className='flex my-1'>
               <p className='w-32'>重量</p>
-              <p>{mac.weight} g</p>
+              <p>{mac.weight} kg</p>
             </div>
           </div>
           <div className='mt-8'>
@@ -284,8 +311,13 @@ const MacShow: NextPage<Props> = ({ id }) => {
           </div>
         </div>
       </div>
-
       <Footer />
+      <AlertMessage
+        open={alertState.open}
+        handleClose={() => setAlertState({ ...alertState, open: false })}
+        type={alertState.type}
+        message={alertState.message}
+      />
     </>
   );
 };
